tool_runner: If offline, use the newest available version of ANTLR4
Previously, if the tool was run while the computer was offline, it would
not work even if any version of ANTLR was already installed previously.
This was because the tool always had to retrieve the latest version from
maven APIs and errored out if that didn't work. Now, we instead try to
look for any already-installed ANTLR4 and use the latest one. This might
provide the user with an out-of-date version of ANTLR4, which is why we
make sure to always report the version to them. However, I consider this
to be better than preventing the user from using the tool altogether
offline.

Signed-off-by: kleines Filmröllchen <filmroellchen@serenityos.org>